Support array on solaris
/bin/sh on solaris is ancient and does not support $() command
subsitution. Additionally, awk on solaris is oawk (old-awk) and it seems
impossible to find a syntax for passing shell variables with embedded
escapes which works on solaris and on linux and the bsds, so I have
resorted to using a case statement to switch to using nawk on solaris.

When iterating the arguments to a function, /bin/sh on solaris also
requires the explicit 'in "$@"' otherwise we get a syntax error.